added clients data validation using regex

+++ src/controller/drivers.controller.ts
@@ -18,6 +18,11 @@ export async function postDrivers(req: Request, res: Response, next: any) {
 
   if (!(name && cluster)) return next(new ErrorException(ErrorCode.WrongInput));
 
+  const regexCheckName = /^[A-Za-z][A-Za-z0-9_]{1,29}$/;
+
+  if (!(regexCheckName.test(name) && /[A-Z]/.test(cluster)))
+    return next(new ErrorException(ErrorCode.WrongInput));
+
   try {
     const results = await driversDB.createOne({ name, cluster });
     res.status(201).json(results);
@@ -33,6 +38,17 @@ export async function putDrivers(req: Request, res: Response, next: any) {
   if (!(name && cluster && newName))
     return next(new ErrorException(ErrorCode.WrongInput));
 
+  const regexCheckName = /^[A-Za-z][A-Za-z0-9_]{1,29}$/;
+
+  if (
+    !(
+      regexCheckName.test(name) &&
+      regexCheckName.test(newName) &&
+      /[A-Z]/.test(cluster)
+    )
+  )
+    return next(new ErrorException(ErrorCode.WrongInput));
+
   try {
     const results = await driversDB.updateOne({ name, newName, cluster });
     res.status(201).json(results);
@@ -44,7 +60,12 @@ export async function putDrivers(req: Request, res: Response, next: any) {
 export async function deleteDrivers(req: Request, res: Response, next: any) {
   const { name } = req.query;
 
+  const regexCheckName = /^[A-Za-z][A-Za-z0-9_]{1,29}$/;
+
   if (!name) return next(new ErrorException(ErrorCode.WrongInput));
+
+  if (!regexCheckName.test(String(name)))
+    return next(new ErrorException(ErrorCode.WrongInput));
 
   try {
     const results = await driversDB.deleteOne(name);
